server/requirements.txt
uvicorn
fastapi
pydantic
bcrypt
mysql-connector-python
python-dotenv
PyJWT